requirements.txt
asgiref==3.9.2
certifi==2025.8.3
charset-normalizer==3.4.3
diff-match-patch==20241021
Django==5.2.7
django-crontab==0.7.1
django-import-export==4.3.10
et_xmlfile==2.0.0
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
whitenoise==6.11.0
